refactor: use mutex instead of locking
